Add Particle and subclasses for sand and mud with physical property calculations
